requirements.txt
# Web scraping
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
playwright==1.40.0
lxml==4.9.3

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Scheduling
schedule==1.2.0

# Utilities
# Note: hashlib, json, re, time, random, datetime, logging, sqlite3, smtplib, email, pathlib, typing, abc are built-in Python modules 